+++ services/app-web/src/pages/QuestionResponse/UploadQuestions/UploadQuestionsForm.tsx
@@ -1,12 +1,11 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import {
-    GridContainer,
     Form as UswdsForm,
     FormGroup,
     ButtonGroup,
 } from '@trussworks/react-uswds'
 import styles from '../QuestionResponse.module.scss'
-import { useNavigate, useOutletContext, useParams } from 'react-router-dom'
+import { useNavigate, useParams } from 'react-router-dom'
 import { useS3 } from '../../../contexts/S3Context'
 import {
     ActionButton,
@@ -15,115 +14,59 @@ import {
     GenericApiErrorBanner,
 } from '../../../components'
 import { useFileUpload } from '../../../hooks/useFileUpload'
-import { ACCEPTED_SUBMISSION_FILE_TYPES } from '../../../components/FileUpload'
+import { ACCEPTED_SUBMISSION_FILE_TYPES, FileItemT } from '../../../components/FileUpload'
 import { PageActionsContainer } from '../../StateSubmission/PageActions'
 import { useErrorSummary } from '../../../hooks/useErrorSummary'
-import {
-    CreateContractQuestionInput,
-    useCreateContractQuestionMutation,
-} from '../../../gen/gqlClient'
-import { SideNavOutletContextType } from '../../SubmissionSideNav/SubmissionSideNav'
-import { usePage } from '../../../contexts/PageContext'
-import { Breadcrumbs } from '../../../components/Breadcrumbs/Breadcrumbs'
-import { createContractQuestionWrapper } from '../../../gqlHelpers/mutationWrappersForUserFriendlyErrors'
-import { RoutesRecord } from '../../../constants'
-import { GenericErrorPage } from '../../Errors/GenericErrorPage'
 
-export const UploadQuestions = () => {
-    // router context
-    // eslint-disable-next-line @typescript-eslint/no-use-before-define
+type UploadQuestionsFormProps =  {
+    handleSubmit: (cleaned: FileItemT[]) => Promise<void>,
+    apiLoading: boolean,
+    apiError: boolean,
+    type: 'contract' | 'rate'
+
+}
+const UploadQuestionsForm = ({handleSubmit, apiError, apiLoading, type}: UploadQuestionsFormProps) => {
     const { division, id } = useParams<{ division: string; id: string }>()
-    const navigate = useNavigate()
-    const { packageName, contract } =
-        useOutletContext<SideNavOutletContextType>()
-
-    // api
-    const [createQuestion, { loading: apiLoading, error: apiError }] =
-        useCreateContractQuestionMutation()
-
-    // page level state
     const [shouldValidate, setShouldValidate] = React.useState(false)
-    const { updateHeading } = usePage()
-    useEffect(() => {
-        updateHeading({ customHeading: `${packageName} Add questions` })
-    }, [packageName, updateHeading])
+    const navigate = useNavigate()
+        const { handleDeleteFile, handleUploadFile, handleScanFile } = useS3()
+        const {
+            hasValidFiles,
+            hasNoFiles,
+            onFileItemsUpdate,
+            fileUploadError,
+            cleanFileItemsBeforeSave,
+        } = useFileUpload(shouldValidate)
+        const { setFocusErrorSummaryHeading, errorSummaryHeadingRef } =
+            useErrorSummary()
 
-    // component specific support
-    const { handleDeleteFile, handleUploadFile, handleScanFile } = useS3()
-    const {
-        hasValidFiles,
-        hasNoFiles,
-        onFileItemsUpdate,
-        fileUploadError,
-        cleanFileItemsBeforeSave,
-    } = useFileUpload(shouldValidate)
-    const { setFocusErrorSummaryHeading, errorSummaryHeadingRef } =
-        useErrorSummary()
+        const uploadComponentID = `${type}-questions-upload`
+        const showFileUploadError = Boolean(shouldValidate && fileUploadError)
+        const fileUploadErrorFocusKey = hasNoFiles
+            ? uploadComponentID
+            : '#file-items-list'
 
-    if (contract.status === 'DRAFT') {
-        return <GenericErrorPage />
-    }
-    const showFileUploadError = Boolean(shouldValidate && fileUploadError)
-    const fileUploadErrorFocusKey = hasNoFiles
-        ? 'questions-upload'
-        : '#file-items-list'
-
-    const handleFormSubmit = async () => {
-        // Currently documents validation happens (outside of the yup schema, which only handles the formik form data)
-        // if there are any errors present in the documents list and we are in a validation state (relevant for Save as Draft) force user to clear validations to continue
-        if (!hasValidFiles || hasNoFiles) {
-            setShouldValidate(true)
-            setFocusErrorSummaryHeading(true)
-            return
-        }
-
-        const cleaned = cleanFileItemsBeforeSave()
-        const questionDocs = cleaned.map((item) => {
-            return {
-                name: item.name,
-                s3URL: item.s3URL as string,
+        const onSubmit = async(e: React.FormEvent<HTMLFormElement>) => {
+                e.preventDefault()
+            // Currently documents validation happens (outside of the yup schema, which only handles the formik form data)
+            // if there are any errors present in the documents list and we are in a validation state (relevant for Save as Draft) force user to clear validations to continue
+            if (!hasValidFiles || hasNoFiles) {
+                setShouldValidate(true)
+                setFocusErrorSummaryHeading(true)
+                return
+            } else {
+                const cleaned = cleanFileItemsBeforeSave()
+                return await handleSubmit(cleaned)
             }
-        })
-
-        const input: CreateContractQuestionInput = {
-            contractID: id as string,
-            documents: questionDocs,
         }
-
-        const createResult = await createContractQuestionWrapper(createQuestion, input)
-
-        if (createResult instanceof Error) {
-            console.info(createResult.message)
-        } else {
-            navigate(`/submissions/${id}/question-and-answers?submit=question`)
-        }
-    }
 
     return (
-        <GridContainer>
-            <Breadcrumbs
-                items={[
-                    {
-                        link: RoutesRecord.DASHBOARD_SUBMISSIONS,
-                        text: 'Dashboard',
-                    },
-                    { link: `/submissions/${id}`, text: packageName },
-                    {
-                        text: 'Add questions',
-                        link: RoutesRecord.SUBMISSIONS_UPLOAD_CONTRACT_QUESTION,
-                    },
-                ]}
-            />
-
-            <UswdsForm
+        <UswdsForm
                 className={styles.formContainer}
-                id="AddQuestionsForm"
+                id={`${type}AddQuestionsForm`}
                 aria-label="Add Questions Form"
                 aria-describedby="form-guidance"
-                onSubmit={async (e) => {
-                    e.preventDefault()
-                    await handleFormSubmit()
-                }}
+                onSubmit={onSubmit}
             >
                 {apiError && <GenericApiErrorBanner />}
                 <fieldset className="usa-fieldset">
@@ -146,8 +89,8 @@ export const UploadQuestions = () => {
 
                     <FormGroup error={showFileUploadError}>
                         <FileUpload
-                            id="questions-upload"
-                            name="questions-upload"
+                            id={uploadComponentID}
+                            name={uploadComponentID}
                             label="Upload questions"
                             aria-required
                             error={showFileUploadError ? fileUploadError : ''}
@@ -204,6 +147,7 @@ export const UploadQuestions = () => {
                     </ButtonGroup>
                 </PageActionsContainer>
             </UswdsForm>
-        </GridContainer>
     )
 }
+
+export {UploadQuestionsForm}
