+++ services/app-web/src/pages/QuestionResponse/UploadQuestions/index.ts
@@ -1,1 +1,1 @@
-export { UploadQuestions } from './UploadQuestions'
+export { UploadContractQuestions } from './UploadContractQuestions'
